% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mci.R
\name{mci_par}
\alias{mci_par}
\title{Calculate marginal contribution feature importance}
\usage{
mci_par(X, y, k = 2)
}
\arguments{
\item{X}{A matrix or dataframe of explanatory features}

\item{y}{A numeric or factor vector}

\item{k}{Assumes soft k-size submodularity}
}
\value{
A numeric vector of feature importance scores
}
\description{
This function calculates the MCI of all features within X in parallel according to random forests accuracy as an evaluation function.
}
