% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_lr.R
\name{preprocess_lr}
\alias{preprocess_lr}
\title{Remove dependencies via linear regression}
\usage{
preprocess_lr(dat, protect)
}
\arguments{
\item{dat}{A dataframe or matrix of data}

\item{protect}{The column number of the protected attribute}
}
\value{
A dataframe or matrix of data with the dependencies between the protected attribute and the rest of the features removed
}
\description{
This function removes the dependencies between the protected attribute and the rest of the features via linear regression.
}
