% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_ot.R
\name{preprocess_ot}
\alias{preprocess_ot}
\title{Remove dependencies via optimal transport}
\usage{
preprocess_ot(dat, protect, n_quan = ceiling(nrow(dat)/150), min_sd = 1e-06)
}
\arguments{
\item{dat}{A dataframe or matrix of data}

\item{protect}{The column number of the protected attribute}

\item{n_quan}{The number of quantiles to use to estimate the conditional CDF (default is ceiling(nrow(dat)/150))}

\item{min_sd}{The minimum standard deviation that data points within a quantile can have before noise is added (default is 1e-6)}
}
\value{
A dataframe or matrix of data with the dependencies between the protected attribute and the rest of the features removed
}
\description{
This function removes the dependencies between the protected attribute and the rest of the features via optimal transport.
}
