% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umfi.R
\name{umfi}
\alias{umfi}
\title{Calculate ultra-marginal feature importance}
\usage{
umfi(X, y, mod_meth = "lr")
}
\arguments{
\item{X}{A matrix or dataframe of explanatory features}

\item{y}{A numeric or factor vector}

\item{mod_meth}{A string indicating if you want to use linear regression ("lr") or optimal transport ("ot") for dependence removal}
}
\value{
A numeric vector of feature importance scores
}
\description{
This function calculates the UMFI of all features within X according to a specified fair representation learning method and random forests accuracy as an evaluation function.
}
